frontend/requirements.txt
PyPDF2==3.0.1
python_docx==1.1.0
streamlit==1.29.0